refactor(zb_io): structural re-organization
