+++ zb_io/src/cellar/mod.rs
@@ -1,0 +1,5 @@
+pub mod link;
+pub mod materialize;
+
+pub use link::{LinkedFile, Linker};
+pub use materialize::{Cellar, CopyStrategy};
