+++ zb_io/src/extraction/mod.rs
@@ -1,0 +1,4 @@
+pub mod extract;
+pub mod patch;
+
+pub use extract::{extract_tarball, extract_tarball_from_reader};
